@@ -33,12 +33,13 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  google_fonts: ^2.3.1
-  font_awesome_flutter: ^10.1.0
-  url_launcher: ^6.0.20
-  firebase_core: ^1.14.1
-  firebase_analytics: ^9.1.5
+  cupertino_icons: ^1.0.5
+  google_fonts: ^5.1.0
+  font_awesome_flutter: ^10.5.0
+  url_launcher: ^6.1.12
+  firebase_core: ^2.14.0
+  firebase_analytics: ^10.4.3
+  flutter_staggered_grid_view: ^0.6.2
 
 dev_dependencies:
   flutter_test:
@@ -50,6 +51,7 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^1.0.0
+  dependency_validator: ^3.2.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -64,7 +66,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/images/image.jpeg
+    - assets/images/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
