@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'Main Page/first_page.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:new_portfolio/Main%20Page/responsive_main_page.dart';
+import 'package:new_portfolio/Projects%20Page/projects.dart';
+import 'package:new_portfolio/Utils/text_theme.dart';
 import 'Services/firebase_config.dart';
 
 void main() async {
@@ -13,39 +14,16 @@ void main() async {
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
-  //final Future<FirebaseApp> _initialization = Firebase.initializeApp();
-  //FirebaseAnalytics analytics = FirebaseAnalytics.instance;
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Chathura',
       theme: ThemeData(
+        textTheme: textTheme,
         primarySwatch: Colors.blue,
       ),
-      home: const HomePage(),
-    );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: "Chathura",
-      theme: ThemeData(
-        textTheme: GoogleFonts.poppinsTextTheme(Theme.of(context).textTheme),
-      ),
-      home: const TopPage(),
+      home: const ProjectsPage(),
     );
   }
 }
